feat(grpc): add client connect

+++ internal/inventory/inventoryEntity.go
@@ -10,10 +10,10 @@ type (
 		entity.Base `bson:"inline"`
 
 		Name             string                    `bson:"name"`
-		PurchasePrice    float64                   `bson:"purchase_price"`
-		PurchaseQuantity float64                   `bson:"purchase_quantity"`
+		PurchasePrice    float32                   `bson:"purchase_price"`
+		PurchaseQuantity float32                   `bson:"purchase_quantity"`
 		PurchaseUnit     *usageUnit.UsageUnitEmbed `bson:"purchase_unit"`
-		YieldPercentage  float64                   `bson:"yield_percentage"`
+		YieldPercentage  float32                   `bson:"yield_percentage"`
 		Remark           string                    `bson:"remark"`
 	}
 )
